+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-
+
+## [1.5.0] - 2026-01-21
+
 ### Added
 
 - Button to delete media in media preview
+- Improved popup (warning, role selection) closing options
+- Category import and export functionality
 
 ### Changed
 
@@ -78,7 +84,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/Braincoke24/QuizBoard/compare/v1.4.1...HEAD
+[unreleased]: https://github.com/Braincoke24/QuizBoard/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/Braincoke24/QuizBoard/compare/v1.4.1...v1.5.0
 [1.4.1]: https://github.com/Braincoke24/QuizBoard/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/Braincoke24/QuizBoard/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/Braincoke24/QuizBoard/compare/v1.2.0...v1.3.0